REFACTOR/ MailComponent: model => component로 패키지 이동
