+++ server/src/main/java/com/rezero/inandout/member/service/MemberServiceImpl.java
@@ -25,7 +25,7 @@ import com.rezero.inandout.member.entity.Member;
 import com.rezero.inandout.member.model.ChangePasswordInput;
 import com.rezero.inandout.member.model.JoinMemberInput;
 import com.rezero.inandout.member.model.LoginMemberInput;
-import com.rezero.inandout.member.model.MailComponent;
+import com.rezero.inandout.member.component.MailComponent;
 import com.rezero.inandout.member.model.MemberDto;
 import com.rezero.inandout.member.model.MemberStatus;
 import com.rezero.inandout.member.model.UpdateMemberInput;
